fix: page load perf by refactoring things out of delayed
